library/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.8
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.9.0
whitenoise==6.6.0
